requirements.txt
flask==3.0.0
pymupdf>=1.24.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.5
google-genai==0.3.0
pageindex==0.1.5
murf>=1.0.0
gunicorn==21.2.0
